+++ classes/Device/AbstractDevice.php
@@ -4,38 +4,11 @@ namespace Device;
 
 use Interfaces\HasNameInterface;
 use Interfaces\HasBrandInterface;
+use Traits\HasNameTrait;
+use Traits\HasBrandTrait;
 
 abstract class AbstractDevice implements HasNameInterface, HasBrandInterface
 {
-    /**
-    * @var string
-    */
-    protected $name;
-
-    /**
-     * @var string
-     */
-    protected $brand;
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(?string $nL): HasNameInterface
-    {
-        $this->name = $nL;
-        return $this;
-    }
-
-    public function getBrand(): ?string
-    {
-        return $this->devices;
-    }
-
-    public function setBrand(?string $b): HasBrandInterface
-    {
-        $this->devices = $b;
-        return $this;
-    }
+    use HasNameTrait;
+    use HasBrandTrait;
 }
